requirements.txt
streamlit
python-docx
groq
pypdf2
python-dotenv